Id field (#589)
* run js in migration

* migration, model

* do not delete, badge

* create

* remove some push ids

* history fields

* tableio

* some table test

* fix test

* getjoined fields

* random pks

* rm log

* api required

* clear all: not users id

* pack drop fkeys

* rm logging

* prefer string

* en

* fix e2e tests

* try catch in migration

+++ packages/saltcorn-data/models/pack.js
@@ -148,7 +148,7 @@ const uninstall_pack = contract(
       if (table) {
         const fields = await table.getFields();
         for (const field of fields) {
-          await field.delete();
+          if (field.is_fkey) await field.delete();
         }
         const triggers = await Trigger.find({ table_id: table.id });
         for (const trigger of triggers) {
